WIP to make the orbit triangle work for any symmetry
The idea is to let each orbit (Direction) compute its dot location, relative to a fundamental
triangle determined by the Symmetry.  I have not worked out a good pattern for this, yet.
As is, the heptagonal antiprism dots are not all falling in the triangle.  Octahedral is pretty good,
but slightly off due to the projection used.