Somewhat failed attempt to improve the mouse tools.
There is some good refactoring here, but the primary goal, to eliminate
the 200 msec lag, is a failure.  Rather, it is eliminated, but the side
effects are some ugly bugs.  Dragging out a strut is now hit or miss,
mostly miss; instead, you get a camera roll.

I may come back to this, but for now I'm discouraged.