WIP to test all edits for correct selection order on undo.
After getting burned by the CommandPolygon defect, we thought it would be
good to have a safety net, so we catch defects related to selection
ordering whenever we create new edit classes.

However, I'm abandoning this testing effort, due to the complexity of
getting it right.  The behavior can vary even for a single class, depending
on how the edit is configured.