Removing AWT dependencies from Controllers, phase 3
Controller no longer uses Graphics, Dimension, MouseEvent, or MouseTool.

Dimension was easy, just replacing with two ints.

For Graphics and MouseEvent, we just switched to Object, so the signature
is clean, and then we cast to MouseEvent in PickingController, and so on.

I'm still having a problem with ClipboardController, but I'll tackle that
in phase 4.