WIP Using a single worker for all viewers
Users want to reuse a single viewer component and switch between designs.
Currently, whenever they change the URL, the worker will fetch and parse
the design again.  The better architecture is to cache the data for each
design by URL, in the worker.

Now, if each worker can cache multiple designs, there is no value to having
multiple workers in the first place.  Furthermore, there is a benefit to
having a single worker, since the global symmetries, etc. can be reused
as in desktop vZome.

Therefore, I'm switching to a single-worker-per-page approach.  Each viewer
component will still have its own store, but each store will have a unique
identifier, and all will send messages to the same worker.  Each message
will be stamped with the storeId, and any responses from the worker must
also have the storeId, so that each response can be directed back to the
appropriate store.